Create extraction and cell vector shape files

create_vectors takes a file path to img format CRW data and a study area shape file and creates extraction point and cell vector shape files.

create_vectors(pathin, aoi)

Arguments

pathin
a character file path to the working directory (see details).

aoi
character representation of the shape file name of the stuady area. No file extension.

Value

two vector files. A point shape file for use in extracting data and a cell boundaries shape file. Both files will be written to the `pathin` and will have an EPSG of 3577.

Details

Takes a file path to img format files (the working directory where `img_data/` is found) and a study area shape file layer name as inputs and creates two vector files.

The first will be used by (extract_daily). It is a point shape file, derived from the centroids of the SST pixel data and can be used to extract SST data for a study area.

The second vector file output is a vectorised representation of the SST raster (cell boundaries) that can be used in further downstream analysis. All outputs have an EPSG of 3577.

The `aoi` shape file will be reprojected to EPSG 3577.

Examples

if (FALSE) {
create_vectors(pathin = ".", aoi = "my_study_area")
}